% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funop.R
\name{funop}
\alias{funop}
\title{Identifies outliers in a numeric vector}
\usage{
funop(x, A = 0, B = 1.5)
}
\arguments{
\item{x}{Numeric vector to inspect for outliers (does not need to be ordered)}

\item{A}{Number of standard deviations beyond the median of \code{x}}

\item{B}{Multiples beyond the median slope of candidate values}
}
\value{
The indices of outliers in \code{x}
}
\description{
FUNOP stands for FUll NOrmal Plot.

The procedure identifies outliers by calculating their slope (\code{z}),
relative to the vector's median.

The procedure ignores values in the middle third of the \emph{ordered}
vector. The remaining values are all candidates for consideration. The
slopes of all candidates are calculated, and the median of their slopes
is used as the primary basis for identifying outliers.

Values whose slope is \code{B} times larger than the median slope are
identified as outliers. The procedure also identifies as outliers any
values whose magnitude is larger than the identified outliers.

However, the procedure will \emph{not} disregard as outliers any values
within \code{A} standard deviations of the vector's median (i.e., not
the median of candidate slopes).
}
\examples{
# funop returns the index of outliers
funop(c(1, 2, 3, 11))
funop(table_1)

# to return the value of outliers, use the function's
# output as the index to the original vector
table_1[funop(table_1)]
}
\references{
Tukey, John W. “The Future of Data Analysis.”
\emph{The Annals of Mathematical Statistics},
\emph{33}(1), 1962, pp 1–67. \emph{JSTOR},
\url{http://www.jstor.org/stable/2237638}.
}
\seealso{
\code{\link[=a_qnorm]{a_qnorm()}}
}
