% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_1.R
\docType{data}
\name{table_1}
\alias{table_1}
\title{Table 1}
\format{
A numeric vector containing 14 elements.
}
\source{
Tukey, John W. "The Future of Data Analysis."
\emph{The Annals of Mathematical Statistics},
\emph{33}(1), 1962, pp 1-67. \emph{JSTOR},
\url{https://www.jstor.org/stable/2237638}.
}
\usage{
table_1
}
\description{
Example data taken from Table 1 of John Tukey's
"Future of Data Analysis."
}
\examples{
table_1
funop(table_1)
}
\keyword{datasets}
