% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_2.R
\docType{data}
\name{table_2}
\alias{table_2}
\title{Table 2}
\format{
An object of class \code{matrix} with 36 rows and 15 columns.
}
\source{
Tukey, John W. "The Future of Data Analysis."
\emph{The Annals of Mathematical Statistics},
\emph{33}(1), 1962, pp 1-67. \emph{JSTOR},
\url{http://www.jstor.org/stable/2237638}.
}
\usage{
table_2
}
\description{
Example data taken from Table 2 of John Tukey's
"Future of Data Analysis."
}
\examples{
{
table_2
funor_funom(table_2)
}
}
\keyword{datasets}
