% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_8.R
\docType{data}
\name{table_8}
\alias{table_8}
\title{Table 8}
\format{
A 36x15 numeric matrix.
}
\source{
Tukey, John W. "The Future of Data Analysis."
\emph{The Annals of Mathematical Statistics},
\emph{33}(1), 1962, pp 1-67. \emph{JSTOR},
\url{http://www.jstor.org/stable/2237638}.
}
\usage{
table_8
}
\description{
Example data taken from Table 8 of John Tukey's
"Future of Data Analysis." Note that this dataset fixes
a typo in the original document: Column 23, row 10 contains
0.100, corrected from the original -0.100.
}
\examples{
table_8
vacuum_cleaner(table_8)
}
\keyword{datasets}
