% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vacuum-package.R
\docType{package}
\name{vacuum-package}
\alias{vacuum}
\alias{vacuum-package}
\title{vacuum: Tukey's Vacuum Cleaner}
\description{
An implementation of three procedures developed by 
    John Tukey: FUNOP, FUNOR-FUNOM, and vacuum cleaner. Combined, they 
    provide a way to identify, treat, and analyze outliers in two-way 
    (i.e., contingency) tables, as described in Tukey's original paper 
    "The Future of Data Analytics".
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Sielinski/vacuum}
  \item Report bugs at \url{https://github.com/Sielinski/vacuum/issues}
}

}
\keyword{internal}
