% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vacuum.R
\name{vacuum}
\alias{vacuum}
\title{Returns the residuals of a two-way table after removing systemic effects}
\usage{
vacuum(x)
}
\arguments{
\item{x}{Two-way table to analyze (must be 3x3 or greater).}
}
\value{
Residuals of \code{x}
}
\description{
To remove systemic effects from values in a contingency table, the
vacuum cleaner uses regression to identify the table's main effect
(dual regression), row effect (deviations of row
regression from dual regression), and column effect (deviations of
column regression from dual regression).

Regression is performed twice: First on the table's original values,
then on the resulting residuals.

The output is a table of residuals "vacuum cleaned" of likely systemic
effects.
}
\examples{
vacuum(table_8)
}
\references{
The techique was orginally developed by John Tukey and published in
"The Future of Data Analytics".
Described in sections 38-42 of Tukey's
paper, starting with "The vacuum cleaner" (pp 49-60), available at
\url{https://www.jstor.org/stable/2237638}.
}
\seealso{
\code{\link[=funop]{funop()}}, \code{\link[=funor_funom]{funor_funom()}}
}
