% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vacuum-package.R, R/vacuum.R
\docType{package}
\name{vacuum-package}
\alias{vacuum}
\alias{vacuum-package}
\title{vacuum: Tukey's Vacuum Cleaner}
\description{
An implementation of three procedures developed by 
    John Tukey: FUNOP (FUll NOrmal Plot), FUNOR-FUNOM 
    (FUll NOrmal Rejection-FUll NOrmal Modification), and vacuum cleaner. 
    Combined, they provide a way to identify, treat, and analyze outliers 
    in two-way (i.e., contingency) tables, as described in his 
    landmark paper "The Future of Data Analysis", Tukey, John W. (1962) 
    <https://www.jstor.org/stable/2237638>.

An implementation of three procedures developed by 
    John Tukey: FUNOP (FUll NOrmal Plot), FUNOR-FUNOM 
    (FUll NOrmal Rejection-FUll NOrmal Modification), and vacuum cleaner. 
    Combined, they provide a way to identify, treat, and analyze outliers 
    in two-way (i.e., contingency) tables, as described in his 
    landmark paper "The Future of Data Analysis", Tukey, John W. (1962) 
    <https://www.jstor.org/stable/2237638>.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Sielinski/vacuum}
  \item Report bugs at \url{https://github.com/Sielinski/vacuum/issues}
}


Useful links:
\itemize{
  \item \url{https://github.com/Sielinski/vacuum}
  \item Report bugs at \url{https://github.com/Sielinski/vacuum/issues}
}

}
\keyword{internal}
